requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
bokeh==3.2.1
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.1
chromadb==0.4.3
click==8.1.6
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.1.0
dataclasses-json==0.5.13
debugpy==1.6.7
decorator==5.1.1
dnspython==2.4.1
email-validator==2.0.0.post2
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.99.1
flatbuffers==23.5.26
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.0
iniconfig==2.0.0
ipykernel==6.25.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.279
langsmith==0.0.33
lark==1.1.7
linkify-it-py==2.0.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.25.1
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.9.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
panel==1.2.1
param==1.13.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
pluggy==1.2.0
posthog==3.0.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pydantic==1.10.12
pydantic_core==2.6.3
Pygments==2.15.1
pypdf==3.13.0
PyPika==0.48.9
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pyviz-comms==2.3.2
PyYAML==6.0.1
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
xyzservices==2023.7.0
yarl==1.9.2
